fix: preserve audit continuation results
