+++ src/benchflow/trajectories/redaction.py
@@ -1,0 +1,383 @@
+"""Canonical stdlib-only secret redaction for trajectory and proxy artifacts."""
+
+from __future__ import annotations
+
+import re
+from typing import Any
+
+# Human-facing redaction categories. Every canonical pattern below is tagged
+# with the kind of secret its rule actually detects, so the upload preview can
+# tell contributors WHAT was masked ("2 API keys, 1 bearer token") instead of
+# only how many values. The taxonomy names the real rules — there is no
+# PII/email detection, so no such category exists.
+REDACTION_CATEGORY_API_KEY = "API key"
+REDACTION_CATEGORY_BEARER_TOKEN = "bearer token"
+REDACTION_CATEGORY_PRIVATE_KEY = "private key block"
+REDACTION_CATEGORY_PASSWORD = "password"
+REDACTION_CATEGORY_URL_CREDENTIAL = "URL credential"
+REDACTION_CATEGORY_CREDENTIAL_FIELD = "credential-bearing field value"
+
+# Quote atom for the header/key-value carriers below: a run of 0-8 optional
+# backslashes followed by an optional quote. This makes every carrier fire on raw
+# text (``"k": "v"``), Python dict-repr (``'k': 'v'``), AND json.dumps-escaped
+# JSON at any nesting depth (``\"k\": \"v\"``, ``\\\"k\\\": …``) — the escaped
+# form still appears when a serialized artifact is redacted as text (the
+# ADP/ATIF exporters, ``results.py``), or when a provider error body embedded in
+# ``error.message`` is itself JSON (#830). (``Trajectory.to_jsonl`` and the ACP
+# writer now redact record *values* before ``json.dumps`` via
+# ``redact_trajectory_obj`` so they can never emit a corrupt escape.) The
+# ``{0,8}`` cap keeps it ReDoS-bounded.
+_ESCQ = r'(?:\\{0,8}["\']|)'
+# Secret value class for the carriers: stops at a quote (plain, or the backslash
+# of an escaped closing quote), whitespace, and the ``,`` ``}`` ``&`` delimiters
+# so a redacted value can never swallow a sibling JSON field or URL query param;
+# excludes ``*`` so an already-redacted value isn't re-matched.
+_SECVAL = r"[^\"'\s,}\\*&]+"
+# Name-prefix class for the env/JSON carriers, LENGTH-CAPPED. An uncapped
+# ``[A-Za-z0-9_]*`` before a required marker backtracks O(n²) on a long
+# alphanumeric run (e.g. a base64 image field), so a benign trajectory could
+# stall the redactor for tens of seconds. Real env-var/header names are short;
+# ``{0,64}`` makes each match attempt O(1) → overall O(n) (#830 ReDoS fix).
+_NAME = r"[A-Za-z0-9_]{0,64}"
+
+_REDACTION_PATTERNS: list[tuple[re.Pattern[str], str, str]] = [
+    # PEM private-key blocks (GCP/Vertex service-account JSON, RSA keys). Redact
+    # the whole armored block incl. the base64 body; the carriers below stop at
+    # the first space and would leave the key material. Lazy + length-capped body
+    # keeps it bounded (#830).
+    (
+        re.compile(
+            r"-----BEGIN [A-Z0-9 ]{0,40}PRIVATE KEY-----"
+            r"[\s\S]{0,8192}?"
+            r"-----END [A-Z0-9 ]{0,40}PRIVATE KEY-----"
+        ),
+        "***REDACTED***",
+        REDACTION_CATEGORY_PRIVATE_KEY,
+    ),
+    # --- Token families: redacted WHOLE (prefix included) so the v0.5 leak audit
+    # greps (``AIzaSy``/``dtn_``…) see no live-key shape (#537/#585). ---
+    # Anthropic: sk-ant-api03-...
+    (
+        re.compile(r"sk-ant-[a-zA-Z0-9_-]{12,}"),
+        "***REDACTED***",
+        REDACTION_CATEGORY_API_KEY,
+    ),
+    # BenchFlow proxy master key (sk-benchflow-<token_urlsafe(24)>), OpenAI service
+    # account (sk-svcacct-), OpenRouter (sk-or-v1-): rare labels that won't appear
+    # in a normal kebab slug, so the hyphen-permitting entropy class is safe here.
+    (
+        re.compile(r"sk-(?:benchflow|svcacct|or-v1)-[A-Za-z0-9_-]{12,}"),
+        "***REDACTED***",
+        REDACTION_CATEGORY_API_KEY,
+    ),
+    # OpenAI org-scoped sk-proj-/sk-admin-: `proj`/`admin` ARE common identifier
+    # words, so a bare hyphen class would redact kebab slugs like
+    # `sk-proj-refactor-auth`. Real keys are high-entropy base64url and always
+    # carry an uppercase char; gate on a lookahead for one so lowercase slugs
+    # survive (#830). Before generic sk- so it wins.
+    (
+        re.compile(
+            r"sk-(?:proj|admin)-(?=[A-Za-z0-9_-]{0,200}[A-Z])[A-Za-z0-9_-]{12,}"
+        ),
+        "***REDACTED***",
+        REDACTION_CATEGORY_API_KEY,
+    ),
+    # OpenAI / generic sk- (alphanumeric only — widening to include `-` would
+    # match common slugs like `task-sk-us-east-1-...`)
+    (re.compile(r"sk-[a-zA-Z0-9]{12,}"), "***REDACTED***", REDACTION_CATEGORY_API_KEY),
+    # Google AI / Gemini: AIzaSy... (≥20 char suffix avoids matching `AIzaSy`
+    # alone). Prefix is redacted too so the audit grep for `AIzaSy` is clean.
+    (
+        re.compile(r"AIzaSy[A-Za-z0-9_-]{20,}"),
+        "***REDACTED***",
+        REDACTION_CATEGORY_API_KEY,
+    ),
+    # AWS access keys: AKIA/ASIA + exactly 16 chars; length anchor avoids
+    # matching English words like "ASIAPACIFIC".
+    (
+        re.compile(r"(?:AKIA|ASIA)[A-Z0-9]{16}(?![A-Z0-9])"),
+        "***REDACTED***",
+        REDACTION_CATEGORY_API_KEY,
+    ),
+    # Daytona SDK tokens: dtn_... — ≥16 char suffix avoids short ids (`dtn_v2`).
+    (
+        re.compile(r"dtn_[A-Za-z0-9_]{16,}"),
+        "***REDACTED***",
+        REDACTION_CATEGORY_API_KEY,
+    ),
+    # GitHub tokens: PATs / OAuth / app / refresh (ghp_/gho_/ghu_/ghs_/ghr_) and
+    # fine-grained PATs (github_pat_...). ≥20 char suffix avoids short slugs.
+    (
+        re.compile(r"(?:ghp|gho|ghu|ghs|ghr)_[A-Za-z0-9]{20,}"),
+        "***REDACTED***",
+        REDACTION_CATEGORY_API_KEY,
+    ),
+    (
+        re.compile(r"github_pat_[A-Za-z0-9_]{20,}"),
+        "***REDACTED***",
+        REDACTION_CATEGORY_API_KEY,
+    ),
+    # Slack tokens: bot/user/app/refresh/legacy (xoxb-/xoxp-/xoxa-/xoxr-/xoxs-).
+    (
+        re.compile(r"xox[baprs]-[A-Za-z0-9-]{10,}"),
+        "***REDACTED***",
+        REDACTION_CATEGORY_API_KEY,
+    ),
+    # Other provider key families with distinctive prefixes — caught here for the
+    # bare-token-in-prose form that no NAME=value carrier sees (a provider
+    # exception that echoes the key). Groq gsk_, xAI xai-, Replicate r8_,
+    # HuggingFace hf_, Fireworks fw_. ≥20-char anchors avoid short ids (#830).
+    (re.compile(r"gsk_[A-Za-z0-9]{20,}"), "***REDACTED***", REDACTION_CATEGORY_API_KEY),
+    (re.compile(r"xai-[A-Za-z0-9]{20,}"), "***REDACTED***", REDACTION_CATEGORY_API_KEY),
+    (re.compile(r"r8_[A-Za-z0-9]{20,}"), "***REDACTED***", REDACTION_CATEGORY_API_KEY),
+    (re.compile(r"hf_[A-Za-z0-9]{20,}"), "***REDACTED***", REDACTION_CATEGORY_API_KEY),
+    (re.compile(r"fw_[A-Za-z0-9]{20,}"), "***REDACTED***", REDACTION_CATEGORY_API_KEY),
+    # JSON Web Tokens (session/bearer creds): three base64url segments split by
+    # dots. The ``eyJ`` prefix is base64url of ``{"`` — a JWT header. The 3rd
+    # (signature) segment requires ≥20 chars (real HS256/RS256 sigs are 43+) so
+    # short dotted base64/method-chains that merely start ``eyJ`` aren't redacted;
+    # a left boundary keeps it from starting mid-identifier. Segment upper bounds
+    # keep it ReDoS-bounded (#830).
+    (
+        re.compile(
+            r"(?<![A-Za-z0-9_-])"
+            r"eyJ[A-Za-z0-9_-]{10,512}\.[A-Za-z0-9_-]{6,512}\.[A-Za-z0-9_-]{20,512}"
+        ),
+        "***REDACTED***",
+        REDACTION_CATEGORY_BEARER_TOKEN,
+    ),
+    # --- URL credentials ---
+    # Credentials in a URL's userinfo (scheme://user:pass@host) for KNOWN
+    # credential-bearing schemes (http/db/cache/mq) — a generic scheme class would
+    # scrub bespoke app deep-links (vscode://a:b@…). Drop the whole userinfo; keep
+    # scheme + host. The username class excludes `:` so it stops deterministically
+    # at the password separator (no catastrophic backtracking), both classes
+    # exclude `?`/`#` so the match can't span into a query string and grab an `@`
+    # from an email param, and the password may contain `@` so an un-encoded `@` in
+    # the password is fully scrubbed (greedy to the last `@`). Length caps keep it
+    # ReDoS-bounded. A plain host:port (no `@`) is untouched (#830).
+    (
+        re.compile(
+            r"((?:https?|ftp|ftps|postgres|postgresql|mysql|mongodb(?:\+srv)?|"
+            r"redis|rediss|amqp|amqps)://)"
+            r"[^\s/?#:@]{0,256}:[^\s/?#]{0,256}@"
+        ),
+        r"\1***REDACTED***@",
+        REDACTION_CATEGORY_URL_CREDENTIAL,
+    ),
+    # Secrets carried as URL query params. Curated EXACT param names (anchored to
+    # `?`/`&` with `=` right after the name) so non-secret params (`?page=`,
+    # `?key=name`, `?author=`, `?auth=basic`) are untouched; the value stops at
+    # `&` so sibling params survive. Bare `key`/`token`/`auth` are intentionally
+    # excluded (too often a non-secret field/mode); so are bare `sig`/`pwd` (a
+    # signature-scheme version / a working-directory path). A real key value under
+    # `?key=` is still caught by its token-family pattern above. Includes AWS
+    # SigV4 / Azure SAS / access-key names. Placed BEFORE the carriers below so the
+    # `*token*` carrier can't over-consume the rest of the query string past `&`.
+    # Like _SECVAL, the value class excludes `*` so an already-redacted
+    # `?access_token=***REDACTED***` isn't re-counted — a re-match here made the
+    # publish-layer stability loop diverge and the server rescan reject (#1008).
+    (
+        re.compile(
+            r"([?&](?:"
+            r"api_?key|access_key|access_token|refresh_token|session_token|"
+            r"session_id|sessionid|client_secret|aws_secret_access_key|account_key|"
+            r"secret|password_hash|password|passwd|"
+            r"signature|x-amz-signature|x-amz-credential|x-amz-security-token|sas"
+            r")=)[^&\s\"'*]+",
+            re.IGNORECASE,
+        ),
+        r"\1***REDACTED***",
+        REDACTION_CATEGORY_URL_CREDENTIAL,
+    ),
+    # Azure SAS `?sig=<token>` — bare `sig` is excluded from the curated list above
+    # (a `?sig=v2` scheme-version flag is a common non-secret), so gate on a
+    # high-entropy value length: a real SAS signature is a 40+ char base64 HMAC,
+    # while a version flag is short (#830). Excludes `*` like the curated params
+    # above so the 14-char `***REDACTED***` marker can never re-match (#1008).
+    (
+        re.compile(r"([?&]sig=)[^&\s\"'*]{16,}", re.IGNORECASE),
+        r"\1***REDACTED***",
+        REDACTION_CATEGORY_URL_CREDENTIAL,
+    ),
+    # --- Header / key-value secret carriers ---
+    # Match `name: value` / `name=value` in raw, JSON, Python dict-repr
+    # (single-quoted), AND json.dumps-escaped (`\"name\": \"value\"`) forms — see
+    # _ESCQ. The name is kept; only the value is dropped (#585/#830). No leading
+    # boundary on the hyphenated header names — the hyphen keeps them from matching
+    # inside underscore variable names.
+    (
+        re.compile(
+            rf"((?:{_ESCQ}(?:x-api-key|x-goog-api-key|api-key){_ESCQ})\s*[:=]\s*{_ESCQ})"
+            rf"{_SECVAL}",
+            re.IGNORECASE,
+        ),
+        r"\1***REDACTED***",
+        REDACTION_CATEGORY_API_KEY,
+    ),
+    # Underscore form `api_key`. No leading boundary: namespaced env dumps like
+    # `GEMINI_API_KEY=secret` / JSON keys such as `"openai_api_key"` must redact
+    # too (#585). The `\1` capture preserves the matched name+separator.
+    (
+        re.compile(
+            rf"({_ESCQ}api_key{_ESCQ}\s*[:=]\s*{_ESCQ}){_SECVAL}", re.IGNORECASE
+        ),
+        r"\1***REDACTED***",
+        REDACTION_CATEGORY_API_KEY,
+    ),
+    # `master_key`/`master-key` and `private_key`/`private-key` carriers — the
+    # BenchFlow proxy master key and GCP/Vertex service-account private-key field
+    # as labelled values (the PEM body itself is scrubbed by the block rule above).
+    (
+        re.compile(
+            rf"({_ESCQ}(?:master|private)[_-]key{_ESCQ}\s*[:=]\s*{_ESCQ}){_SECVAL}",
+            re.IGNORECASE,
+        ),
+        r"\1***REDACTED***",
+        REDACTION_CATEGORY_CREDENTIAL_FIELD,
+    ),
+    # Bearer-token env vars whose NAME has BEARER_TOKEN in the MIDDLE, e.g.
+    # `AWS_BEARER_TOKEN_BEDROCK=` — the secret-suffix rule below anchors its marker
+    # at the name end, so a TOKEN in the middle slips through. `BEARER_TOKEN` is an
+    # unambiguous secret signal, safe as a substring (#830). Names length-capped.
+    (
+        re.compile(
+            rf"(?<![A-Za-z0-9])"
+            rf"({_ESCQ}{_NAME}BEARER_TOKEN{_NAME}{_ESCQ}\s*[:=]\s*{_ESCQ}){_SECVAL}",
+            re.IGNORECASE,
+        ),
+        r"\1***REDACTED***",
+        REDACTION_CATEGORY_BEARER_TOKEN,
+    ),
+    # Generic secret-NAME-suffix carriers: names ending in TOKEN/SECRET/PASSWORD or
+    # a specific *secret* key suffix (ACCESS_KEY/SECRET_KEY/ACCOUNT_KEY/…). The KEY
+    # forms are spelled out (NOT a bare `_KEY`) so common non-secret identifiers
+    # like `primary_key`/`foreign_key`/`sort_key` are NOT redacted, while
+    # `AWS_SECRET_ACCESS_KEY`/`AZURE_STORAGE_ACCOUNT_KEY` are. TOKEN/SECRET at the
+    # name end stay safe from `TOKENIZER`/`SECRETARY` (marker not at the end).
+    (
+        re.compile(
+            rf"(?<![A-Za-z0-9])({_ESCQ}{_NAME}"
+            r"(?:TOKEN|SECRET|PASSWORD|PASSWD|"
+            r"ACCESS_KEY|SECRET_KEY|ACCOUNT_KEY|PRIVATE_KEY|ENCRYPTION_KEY|"
+            r"CREDENTIALS?)"
+            rf"{_ESCQ}\s*[:=]\s*{_ESCQ}){_SECVAL}",
+            re.IGNORECASE,
+        ),
+        r"\1***REDACTED***",
+        REDACTION_CATEGORY_CREDENTIAL_FIELD,
+    ),
+    # authorization header WITH a scheme (Bearer/Token/Basic/…): keep scheme.
+    (
+        re.compile(
+            rf"(?<![A-Za-z0-9_-])({_ESCQ}authorization{_ESCQ}\s*[:=]\s*{_ESCQ}"
+            rf"(?:Bearer|Token|Basic|ApiKey)\s+){_SECVAL}",
+            re.IGNORECASE,
+        ),
+        r"\1***REDACTED***",
+        REDACTION_CATEGORY_BEARER_TOKEN,
+    ),
+    # authorization header with a bare value (no recognized scheme). The negative
+    # lookahead skips scheme-prefixed values already handled above, so
+    # `Bearer <tok>` isn't double-redacted into `***REDACTED*** ***...`.
+    (
+        re.compile(
+            rf"(?<![A-Za-z0-9_-])({_ESCQ}authorization{_ESCQ}\s*[:=]\s*{_ESCQ})"
+            rf"(?!(?:Bearer|Token|Basic|ApiKey)\b|[rRuUbBfF]{{1,3}}[\"'])"
+            rf"{_SECVAL}",
+            re.IGNORECASE,
+        ),
+        r"\1***REDACTED***",
+        REDACTION_CATEGORY_BEARER_TOKEN,
+    ),
+]
+
+
+def redact_trajectory_text(text: str) -> str:
+    """Apply all secret-redaction patterns to *text*.
+
+    Token families (Anthropic/OpenAI/Google/AWS/Daytona/GitHub/Slack) are
+    redacted whole, prefix included, so the secret-leak audit greps see no
+    live-key shape. Header/key-value carriers keep the field name but drop the
+    value, in both JSON (``"x-api-key": "v"``) and raw-text (``x-api-key: v``)
+    forms; this includes generic ``*TOKEN*``/``*SECRET*`` carriers so a
+    ``GITHUB_TOKEN=...`` env dump is scrubbed even without a known prefix.
+    """
+    return redact_trajectory_text_with_count(text)[0]
+
+
+def redact_trajectory_text_with_count(text: str) -> tuple[str, int]:
+    """Apply the canonical patterns and report how many matches were replaced."""
+    text, categories = redact_trajectory_text_with_categories(text)
+    return text, sum(categories.values())
+
+
+def redact_trajectory_text_with_categories(text: str) -> tuple[str, dict[str, int]]:
+    """Apply the canonical patterns and report replacements per category.
+
+    Categories name the kind of secret each rule actually detects (API key,
+    bearer token, private key block, URL credential, credential-bearing field
+    value) so upload previews can itemize what was masked without inventing
+    detection that does not exist. The total replacement count is the sum of
+    the returned values.
+    """
+    categories: dict[str, int] = {}
+    for pattern, replacement, category in _REDACTION_PATTERNS:
+        text, count = pattern.subn(replacement, text)
+        if count:
+            categories[category] = categories.get(category, 0) + count
+    return text, categories
+
+
+def redact_trajectory_obj(obj: Any) -> Any:
+    """Recursively redact secrets in the string leaves of a JSON-serializable value.
+
+    Prefer this over running :func:`redact_trajectory_text` on an
+    already-``json.dumps``-ed record. Redacting the serialized *text* can split a
+    ``\\`` escape — a secret sitting next to one leaves a lone ``\\X`` — which
+    corrupts the JSON so the trajectory file no longer parses and
+    ``bench train convert`` fails. Redacting the *values* before serializing keeps
+    escaping valid while preserving coverage: free-text leaves are scrubbed by the
+    token-family and inline ``name: value`` carriers, and structured fields (e.g.
+    a ``{"x-api-key": "<prefixless>"}`` header) are scrubbed in their key context
+    by :func:`_redact_field` so the carriers that key off a field name still fire.
+    """
+    if isinstance(obj, dict):
+        return {key: _redact_field(key, value) for key, value in obj.items()}
+    if isinstance(obj, list):
+        return [redact_trajectory_obj(item) for item in obj]
+    if isinstance(obj, str):
+        return redact_trajectory_text(obj)
+    return obj
+
+
+def _redact_field(key: Any, value: Any) -> Any:
+    """Redact a dict field's *value*, keeping the field name as carrier context.
+
+    The ``name: value`` carriers (``x-api-key``, ``api_key``, ``authorization``,
+    ``*_TOKEN`` …) strip *prefixless* secret values only when they can see the
+    field name; in a parsed object the name and value are separate. So redact the
+    value inside a ``{"<key>": "<value>"}`` probe and read it back. The probe
+    round-trips through json so escaping stays valid; if redacting the serialized
+    probe would corrupt it (the escape hazard this module guards), fall back to
+    redacting the bare value, which still catches the prefixed token families.
+    """
+    if isinstance(value, (dict, list)):
+        return redact_trajectory_obj(value)
+    if not isinstance(value, str):
+        return value
+    if isinstance(key, str):
+        import json
+
+        probe = json.dumps({key: value})
+        redacted = redact_trajectory_text(probe)
+        if redacted == probe:
+            return value
+        try:
+            restored = json.loads(redacted)
+        except ValueError:
+            pass
+        else:
+            if isinstance(restored, dict) and isinstance(restored.get(key), str):
+                return restored[key]
+    return redact_trajectory_text(value)
